Extracted module `BasicMenu`.
Still poking at it; covered by specs for `BuildMenuFor::RegisteredUser`.

RSpec: 1024 examples, 0 failures, 1 pending
RuboCop: 293 files inspected, no offenses detected
Coverage: 3132 / 3146 LOC (99.55%) covered.
