+++ app/helpers/application_helper/build_menu_for/registered_user.rb
@@ -1,19 +1,19 @@
 
 # require 'contracts'
 
-require_relative 'menu_for_details_selector'
+require_relative 'basic_menu'
 
 module ApplicationHelper
   # Module containing `ApplicationHelper#build_menu_for` and support classes.
   module BuildMenuFor
     class RegisteredUser
       include Contracts
+      include BasicMenu
 
       Contract ViewHelper, Or[:navbar, :sidebar],
                RespondTo[:name] => RegisteredUser
       def initialize(h, which, current_user)
-        extend MenuForDetailsSelector.new.select(which)
-        @h = h
+        init_basic_menu h, which
         @current_user = current_user
         self
       end
@@ -21,11 +21,6 @@ module ApplicationHelper
       Contract None => String
       def markup
         build_container do
-          build_item_for 'Home', href: h.root_path
-          build_item_for 'All members', href: h.users_path
-          build_separator_item
-          build_item_for 'New Post', href: h.new_post_path
-          build_separator_item
           build_item_for 'View your profile', href: user_profile_path
           build_separator_item
           build_item_for 'Log out', logout_params
@@ -34,33 +29,7 @@ module ApplicationHelper
 
       private
 
-      attr_reader :current_user, :h
-
-      Contract Proc => String
-      def build_container(&block)
-        h.content_tag :ul, class: container_classes do
-          instance_eval(&block)
-        end
-      end
-
-      Contract String, Maybe[Hash] => String
-      def build_item_for(text, attrs = {})
-        item = h.content_tag :li do
-          link = h.content_tag :a, attrs do
-            text
-          end
-          h.concat link
-        end
-        h.concat item
-      end
-
-      Contract None => String
-      def build_separator_item
-        item = h.content_tag :li, separator_attrs do
-          HTMLEntities.new.decode '&nbsp;'
-        end
-        h.concat item
-      end
+      attr_reader :current_user
 
       Contract None => Hash
       def logout_params
